+++ python/lsst/ts/IntegrationTests/auxtel_housekeeping.py
@@ -38,17 +38,17 @@ class AuxTelHousekeeping(BaseScript):
 
     index: int = 2
     configs: tuple = (
-        registry["atdome_housekeeping"],
         [],
-        [],
-        registry["atdome_housekeeping"],
+        registry["atdome_home"],
+        registry["atdome_park"],
+        registry["atptg_park"],
+        registry["atptg_stop_tracking"],
         registry["atmcs_housekeeping"],
-        registry["atspectrograph_housekeeping"],
     )
     scripts: list = [
-        ("run_command.py", BaseScript.is_standard),
-        ("auxtel/shutdown.py", BaseScript.is_standard),
         ("auxtel/enable_atcs.py", BaseScript.is_standard),
+        ("run_command.py", BaseScript.is_standard),
+        ("run_command.py", BaseScript.is_standard),
         ("run_command.py", BaseScript.is_standard),
         ("run_command.py", BaseScript.is_standard),
         ("run_command.py", BaseScript.is_standard),